django_web_scraping/requirements.txt
amqp==2.6.1
asgiref==3.2.10
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.1.1
defusedxml==0.7.0rc1
Django==3.1.2
django-celery-results==1.2.1
django-crispy-forms==1.9.2
idna==2.10
importlib-metadata==2.0.0
kombu==4.6.11
lxml==4.5.2
oauthlib==3.1.0
Pillow==7.2.0
prompt-toolkit==3.0.8
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
zipp==3.3.0
